backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0
Flask-Compress==1.14
Werkzeug==3.0.1

# Security dependencies
PyJWT==2.8.0
cryptography==41.0.7
python-magic==0.4.27
bcrypt==4.1.2

# Batfish client
pybatfish==2024.7.22.1569

# Data handling
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
click==8.1.7

# Production server
gunicorn==21.2.0
gevent==23.9.1

# Redis for caching/sessions (production)
redis==5.0.1
hiredis==2.2.3

# Database (optional, for future use)
SQLAlchemy==2.0.23
alembic==1.13.0

# Monitoring and logging
python-json-logger==2.0.7
sentry-sdk[flask]==1.39.1

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Security scanning and validation
bandit==1.7.5
safety==3.0.1

# Development dependencies
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
pre-commit==3.6.0